decompose feature
