+++ kswift-gradle-plugin/src/main/kotlin/dev/icerock/moko/kswift/plugin/feature/associatedenum/NamingMode.kt
@@ -1,0 +1,3 @@
+package dev.icerock.moko.kswift.plugin.feature.associatedenum
+
+internal enum class NamingMode { KOTLIN, KOTLIN_NO_STRING, SWIFT, OBJC }
